renaming mongoose connector to connect.js
