+++ lib/connect.js
