+++ test/e2e/app.test.js
@@ -1,5 +1,5 @@
 require('dotenv').config();
-require('../../lib/mongoose-connector')();
+require('../../lib/connect')();
 const mongoose = require('mongoose');
 const request = require('supertest');
 const app = require('../../lib/app');
